Run the function if called as script; named changed accordingly
Changed the name to just start-remote-env since it might be started
non-interactively (i.e. via tmux or such). For such a case, I added
code which merely wraps the defined function.

Though it may be started non-interactively, it should still be auto-
sourced for interactive shells, so I updated the auto-start-remote-env
links for bash and zsh to the new filename.

Since the wrapping now takes place in the file itself, I'm removing
the simple wrapper file.